refactor: move font adapted text component to entities directory
#48
