+++ src/entities/fonts/TextRegular.tsx
