+++ src/entities/fonts/TextThin.tsx
